[#136048] Extract second mailer
