+++ app/views/purchase_notifier/order_receipt.html.haml
