+++ app/views/purchase_notifier/order_receipt.text.haml
