+++ app/views/purchase_notifier/product_order_notification.html.haml
