+++ app/views/purchase_notifier/product_order_notification.text.haml
