+++ config/locales/en.notifier.yml
@@ -8,15 +8,6 @@ en:
         intro: "Congratulations, %{first_name}. A !app_name! account has been created for you. You may log in at %{login_url}."
         no_password: You may log in with your username and password.
         subject: "Welcome to !app_name!"
-      product_order_notification:
-        subject: "%{product} Order Notification"
-      order_notification:
-        subject: "!app_name! Order Notification"
-      order_receipt:
-        subject: "!app_name! Order Receipt"
-        intro: "Thank you for your order on the !app_name! website."
-        outro: "All prices are estimates. Actual cost is assigned when the order is complete."
-        total: Total
       order_status_changed_to_approved:
         subject: "Order #%{order_detail} has been approved"
         header: "The following order has been approved."
